feat: 수정/삭제 시, 비밀번호 비교 로직 추가

+++ src/main/java/com/wisdom/blogwithoutlogin/service/impl/BlogServiceImpl.java
@@ -34,20 +34,25 @@ public class BlogServiceImpl implements BlogService {
         List<Blog> blogs = blogRepository.findAll(Sort.by(Sort.Direction.DESC, "createdDate"));
         return blogs.stream().map(m -> new BlogResponseDto(m.getName(), m.getTitle(), m.getContents(), m.getCreatedDate())).collect(Collectors.toList());
     }
-    
+
     @Override
     public BlogResponseDetailDto getBlogById(long id) {
-        Blog blog = blogRepository.findById(id).orElseThrow(() ->
-                new ResourceNotFoundException("Blog", "Id", id));
+        Blog blog = exists(id);
         return new BlogResponseDetailDto(blog.getName(), blog.getTitle(), blog.getContents(), blog.getCreatedDate());
     }
 
     @Override
-    public Blog updateBlog(long id, BlogUpdateRequestDto blogSaveRequestDto) {
-        Blog existingBlog = blogRepository.findById(id).orElseThrow(() ->
-                new ResourceNotFoundException("Blog", "Id", id));
+    public Blog updateBlog(long id, BlogUpdateRequestDto blogUpdateRequestDto) {
 
-        Blog blog = blogSaveRequestDto.toEntity();
+        Blog existingBlog = exists(id);
+
+        String inputPassword = blogUpdateRequestDto.getPassword();
+        boolean validatePassword = checkPassword(id, inputPassword);
+        if (!validatePassword) {
+            throw new RuntimeException("비밀번호를 다시 입력해주세요.");
+        }
+
+        Blog blog = blogUpdateRequestDto.toEntity();
         existingBlog.setTitle(blog.getTitle());
         existingBlog.setContents(blog.getContents());
         blogRepository.save(existingBlog);
@@ -55,9 +60,24 @@ public class BlogServiceImpl implements BlogService {
     }
 
     @Override
-    public void deleteBlogById(long id) {
-        blogRepository.findById(id).orElseThrow(() ->
-                new ResourceNotFoundException("Blog", "Id", id));
+    public void deleteBlogById(String password, long id) {
+        boolean validatePassword = checkPassword(id, password);
+        System.out.println("validatePassword = " + validatePassword);
+
+        if (!validatePassword) {
+            throw new RuntimeException("비밀번호를 다시 입력해주세요.");
+        }
         blogRepository.deleteById(id);
+    }
+
+    @Override
+    public boolean checkPassword(long id, String password) {
+        Blog blog = exists(id);
+        return password.equals(blog.getPassword());
+    }
+
+    private Blog exists(long id) {
+        return  blogRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Blog", "Id", id));
     }
 }
